+++ src/Twilio/Rest/Api/V2010/Account/Conference/ParticipantOptions.php
@@ -36,7 +36,7 @@ abstract class ParticipantOptions
      * @param bool $earlyMedia Whether to allow an agent to hear the state of the outbound call, including ringing or disconnect messages. Can be: `true` or `false` and defaults to `true`.
      * @param int $maxParticipants The maximum number of participants in the conference. Can be a positive integer from `2` to `250`. The default value is `250`.
      * @param string $conferenceRecord Whether to record the conference the participant is joining. Can be: `true`, `false`, `record-from-start`, and `do-not-record`. The default value is `false`.
-     * @param string $conferenceTrim Whether to trim leading and trailing silence from your recorded conference audio files. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
+     * @param string $conferenceTrim Whether to trim leading and trailing silence from the conference recording. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
      * @param string $conferenceStatusCallback The URL we should call using the `conference_status_callback_method` when the conference events in `conference_status_callback_event` occur. Only the value set by the first participant to join the conference is used. Subsequent `conference_status_callback` values are ignored.
      * @param string $conferenceStatusCallbackMethod The HTTP method we should use to call `conference_status_callback`. Can be: `GET` or `POST` and defaults to `POST`.
      * @param string[] $conferenceStatusCallbackEvent The conference state changes that should generate a call to `conference_status_callback`. Can be: `start`, `end`, `join`, `leave`, `mute`, `hold`, `modify`, `speaker`, and `announcement`. Separate multiple values with a space. Defaults to `start end`.
@@ -65,6 +65,7 @@ abstract class ParticipantOptions
      * @param int $machineDetectionSilenceTimeout The number of milliseconds of initial silence after which an `unknown` AnsweredBy result will be returned. Possible Values: 2000-10000. Default: 5000.
      * @param string $amdStatusCallback The URL that we should call using the `amd_status_callback_method` to notify customer application whether the call was answered by human, machine or fax.
      * @param string $amdStatusCallbackMethod The HTTP method we should use when calling the `amd_status_callback` URL. Can be: `GET` or `POST` and the default is `POST`.
+     * @param string $trim Whether to trim any leading and trailing silence from the participant recording. Can be: `trim-silence` or `do-not-trim` and the default is `trim-silence`.
      * @return CreateParticipantOptions Options builder
      */
     public static function create(
@@ -112,7 +113,8 @@ abstract class ParticipantOptions
         int $machineDetectionSpeechEndThreshold = Values::INT_NONE,
         int $machineDetectionSilenceTimeout = Values::INT_NONE,
         string $amdStatusCallback = Values::NONE,
-        string $amdStatusCallbackMethod = Values::NONE
+        string $amdStatusCallbackMethod = Values::NONE,
+        string $trim = Values::NONE
 
     ): CreateParticipantOptions
     {
@@ -160,7 +162,8 @@ abstract class ParticipantOptions
             $machineDetectionSpeechEndThreshold,
             $machineDetectionSilenceTimeout,
             $amdStatusCallback,
-            $amdStatusCallbackMethod
+            $amdStatusCallbackMethod,
+            $trim
         );
     }
 
@@ -255,7 +258,7 @@ class CreateParticipantOptions extends Options
      * @param bool $earlyMedia Whether to allow an agent to hear the state of the outbound call, including ringing or disconnect messages. Can be: `true` or `false` and defaults to `true`.
      * @param int $maxParticipants The maximum number of participants in the conference. Can be a positive integer from `2` to `250`. The default value is `250`.
      * @param string $conferenceRecord Whether to record the conference the participant is joining. Can be: `true`, `false`, `record-from-start`, and `do-not-record`. The default value is `false`.
-     * @param string $conferenceTrim Whether to trim leading and trailing silence from your recorded conference audio files. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
+     * @param string $conferenceTrim Whether to trim leading and trailing silence from the conference recording. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
      * @param string $conferenceStatusCallback The URL we should call using the `conference_status_callback_method` when the conference events in `conference_status_callback_event` occur. Only the value set by the first participant to join the conference is used. Subsequent `conference_status_callback` values are ignored.
      * @param string $conferenceStatusCallbackMethod The HTTP method we should use to call `conference_status_callback`. Can be: `GET` or `POST` and defaults to `POST`.
      * @param string[] $conferenceStatusCallbackEvent The conference state changes that should generate a call to `conference_status_callback`. Can be: `start`, `end`, `join`, `leave`, `mute`, `hold`, `modify`, `speaker`, and `announcement`. Separate multiple values with a space. Defaults to `start end`.
@@ -284,6 +287,7 @@ class CreateParticipantOptions extends Options
      * @param int $machineDetectionSilenceTimeout The number of milliseconds of initial silence after which an `unknown` AnsweredBy result will be returned. Possible Values: 2000-10000. Default: 5000.
      * @param string $amdStatusCallback The URL that we should call using the `amd_status_callback_method` to notify customer application whether the call was answered by human, machine or fax.
      * @param string $amdStatusCallbackMethod The HTTP method we should use when calling the `amd_status_callback` URL. Can be: `GET` or `POST` and the default is `POST`.
+     * @param string $trim Whether to trim any leading and trailing silence from the participant recording. Can be: `trim-silence` or `do-not-trim` and the default is `trim-silence`.
      */
     public function __construct(
         
@@ -330,7 +334,8 @@ class CreateParticipantOptions extends Options
         int $machineDetectionSpeechEndThreshold = Values::INT_NONE,
         int $machineDetectionSilenceTimeout = Values::INT_NONE,
         string $amdStatusCallback = Values::NONE,
-        string $amdStatusCallbackMethod = Values::NONE
+        string $amdStatusCallbackMethod = Values::NONE,
+        string $trim = Values::NONE
 
     ) {
         $this->options['statusCallback'] = $statusCallback;
@@ -377,6 +382,7 @@ class CreateParticipantOptions extends Options
         $this->options['machineDetectionSilenceTimeout'] = $machineDetectionSilenceTimeout;
         $this->options['amdStatusCallback'] = $amdStatusCallback;
         $this->options['amdStatusCallbackMethod'] = $amdStatusCallbackMethod;
+        $this->options['trim'] = $trim;
     }
 
     /**
@@ -560,9 +566,9 @@ class CreateParticipantOptions extends Options
     }
 
     /**
-     * Whether to trim leading and trailing silence from your recorded conference audio files. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
+     * Whether to trim leading and trailing silence from the conference recording. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
      *
-     * @param string $conferenceTrim Whether to trim leading and trailing silence from your recorded conference audio files. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
+     * @param string $conferenceTrim Whether to trim leading and trailing silence from the conference recording. Can be: `trim-silence` or `do-not-trim` and defaults to `trim-silence`.
      * @return $this Fluent Builder
      */
     public function setConferenceTrim(string $conferenceTrim): self
@@ -904,6 +910,18 @@ class CreateParticipantOptions extends Options
     public function setAmdStatusCallbackMethod(string $amdStatusCallbackMethod): self
     {
         $this->options['amdStatusCallbackMethod'] = $amdStatusCallbackMethod;
+        return $this;
+    }
+
+    /**
+     * Whether to trim any leading and trailing silence from the participant recording. Can be: `trim-silence` or `do-not-trim` and the default is `trim-silence`.
+     *
+     * @param string $trim Whether to trim any leading and trailing silence from the participant recording. Can be: `trim-silence` or `do-not-trim` and the default is `trim-silence`.
+     * @return $this Fluent Builder
+     */
+    public function setTrim(string $trim): self
+    {
+        $this->options['trim'] = $trim;
         return $this;
     }