+++ src/Twilio/Rest/Intelligence.php
@@ -1,0 +1,5 @@
+<?php
+namespace Twilio\Rest;
+
+class Intelligence extends IntelligenceBase {
+}